fix: Add waiting close channels to channels list (#74)
* Add waiting close channels to channels list

Channels in the waiting close state are stuck in the mempool awaiting confimation. This is why the local and remote balance is unconfirmed.

* Remove "Closing in: " panel during waiting close state.

Its impossible to give a good estimate for when the channel will close before the transaction is committed to the blockchain. The panel will reappear once the transaction is committed to a block